refactor: metrics_{buffer,router} -> metrics_pipeline
consolidate these related packages into a single package.
