+++ src/stackdriver-nozzle/app/builder.go
@@ -10,8 +10,7 @@ import (
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/cloudfoundry"
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/config"
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/heartbeat"
-	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/metrics_buffer"
-	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/metrics_router"
+	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/metrics_pipeline"
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/nozzle"
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/stackdriver"
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/version"
@@ -107,7 +106,7 @@ func (a *App) newConsumer(ctx context.Context) (*nozzle.Nozzle, error) {
 	// Destination for metrics
 	metricAdapter := a.newMetricAdapter()
 	// Routes metrics to Stackdriver Logging/Stackdriver Monitoring
-	metricRouter := metrics_router.NewMetricsRouter(metricAdapter, metricEvents, logAdapter, logEvents)
+	metricRouter := metrics_pipeline.NewRouter(metricAdapter, metricEvents, logAdapter, logEvents)
 	// Handles and translates Firehose events. Performs buffering/culling.
 	metricSink := a.newMetricSink(ctx, metricRouter)
 	// Filter Firehose events to what the user selects
@@ -154,7 +153,7 @@ func (a *App) newMetricAdapter() stackdriver.MetricAdapter {
 }
 
 func (a *App) newMetricSink(ctx context.Context, metricAdapter stackdriver.MetricAdapter) nozzle.Sink {
-	metricBuffer, errs := metrics_buffer.NewAutoCulledMetricsBuffer(ctx, a.logger, time.Duration(a.c.MetricsBufferDuration)*time.Second, a.c.MetricsBufferSize, metricAdapter)
+	metricBuffer, errs := metrics_pipeline.NewAutoCulledMetricsBuffer(ctx, a.logger, time.Duration(a.c.MetricsBufferDuration)*time.Second, a.c.MetricsBufferSize, metricAdapter)
 	a.bufferEmpty = metricBuffer.IsEmpty
 	go func() {
 		for err := range errs {
