--- src/stackdriver-nozzle/metrics_buffer/metrics_buffer_suite_test.go
@@ -1,13 +1,0 @@
-package metrics_buffer_test
-
-import (
-	. "github.com/onsi/ginkgo"
-	. "github.com/onsi/gomega"
-
-	"testing"
-)
-
-func TestMetricsBuffer(t *testing.T) {
-	RegisterFailHandler(Fail)
-	RunSpecs(t, "MetricsBuffer Suite")
-}
