+++ src/stackdriver-nozzle/metrics_pipeline/auto_culled_metrics_buffer.go
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package metrics_buffer
+package metrics_pipeline
 
 import (
 	"context"
