+++ src/stackdriver-nozzle/metrics_pipeline/router.go
@@ -1,4 +1,4 @@
-package metrics_router
+package metrics_pipeline
 
 import (
 	"github.com/cloudfoundry-community/stackdriver-tools/src/stackdriver-nozzle/messages"
@@ -6,17 +6,17 @@ import (
 	"github.com/cloudfoundry/sonde-go/events"
 )
 
-type metricRouter struct {
+type router struct {
 	metricAdapter stackdriver.MetricAdapter
 	logAdapter    stackdriver.LogAdapter
 	logEvents     map[events.Envelope_EventType]bool
 	metricEvents  map[events.Envelope_EventType]bool
 }
 
-// NewMetricRouter provides a MetricAdapter that routes a given metric to
+// NewRouter provides a MetricAdapter that routes a given metric to
 // Stackdriver Logging and Stackdriver Monitoring based on configuration
-func NewMetricsRouter(metricAdapter stackdriver.MetricAdapter, metricEvents []events.Envelope_EventType, logAdapter stackdriver.LogAdapter, logEvents []events.Envelope_EventType) stackdriver.MetricAdapter {
-	r := &metricRouter{metricAdapter: metricAdapter, logAdapter: logAdapter}
+func NewRouter(metricAdapter stackdriver.MetricAdapter, metricEvents []events.Envelope_EventType, logAdapter stackdriver.LogAdapter, logEvents []events.Envelope_EventType) stackdriver.MetricAdapter {
+	r := &router{metricAdapter: metricAdapter, logAdapter: logAdapter}
 
 	r.metricEvents = make(map[events.Envelope_EventType]bool)
 	for _, e := range metricEvents {
@@ -31,7 +31,7 @@ func NewMetricsRouter(metricAdapter stackdriver.MetricAdapter, metricEvents []ev
 	return r
 }
 
-func (r *metricRouter) PostMetrics(metrics []messages.Metric) error {
+func (r *router) PostMetrics(metrics []messages.Metric) error {
 	for _, metric := range metrics {
 		if r.metricEvents[metric.Type] {
 			// TODO: seems strange to re-package this as a new slice
